+++ src/valency_anndata/datasets/__init__.py
@@ -1,12 +1,16 @@
 from .polis import load, translate_statements
+from ._load_american_assembly import american_assembly
 from ._load_aufstehen import aufstehen
+from ._load_bg2050 import bg2050
 from ._load_chile_protest import chile_protest
 from ._load_cuba_protest import cuba_protest
 from ._load_klimarat import klimarat
 
 __all__ = [
     "load",
+    "american_assembly",
     "aufstehen",
+    "bg2050",
     "chile_protest",
     "cuba_protest",
     "klimarat",